+++ src/app/features/home/custom-camera/custom-camera.page.ts
@@ -2,9 +2,10 @@
 import { Location } from '@angular/common';
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { PluginListenerHandle } from '@capacitor/core';
+import { Capacitor, PluginListenerHandle } from '@capacitor/core';
 import { UntilDestroy } from '@ngneat/until-destroy';
 import { CaptureResult, PreviewCamera } from '@numbersprotocol/preview-camera';
+import { BehaviorSubject } from 'rxjs';
 import { ErrorService } from '../../../shared/error/error.service';
 import { UserGuideService } from '../../../shared/user-guide/user-guide.service';
 import { GoProBluetoothService } from '../../settings/go-pro/services/go-pro-bluetooth.service';
@@ -32,6 +33,13 @@ export class CustomCameraPage implements OnInit, OnDestroy {
   scaleDownAnimationAfterDelay = 230;
 
   curSessionCaptureMediaItems: CustomCameraMediaItem[] = [];
+
+  mode$ = new BehaviorSubject<'capture' | 'pre-publish'>('capture');
+
+  curCaptureFilePath?: string;
+  curCaptureMimeType?: 'image/jpeg' | 'video/mp4';
+  curCaptureType?: 'image' | 'video' = 'image';
+  curCaptureSrc?: string;
 
   readonly lastConnectedGoProDevice$ =
     this.goProBluetoothService.lastConnectedDevice$;
@@ -85,7 +93,18 @@ export class CustomCameraPage implements OnInit, OnDestroy {
     if (data.errorMessage) {
       await this.errorService.toastError$(data.errorMessage).toPromise();
     } else if (data.filePath) {
-      this.customCameraService.uploadToCapture(data.filePath, type);
+      const filePath = data.filePath;
+
+      let mimeType: 'image/jpeg' | 'video/mp4' = 'image/jpeg';
+      if (type === 'video') mimeType = 'video/mp4';
+
+      this.curCaptureFilePath = filePath;
+      this.curCaptureMimeType = mimeType;
+      this.curCaptureType = type;
+      this.curCaptureSrc = Capacitor.convertFileSrc(filePath);
+      this.mode$.next('pre-publish');
+
+      this.stopPreviewCamera();
     }
   }
 
@@ -131,6 +150,23 @@ export class CustomCameraPage implements OnInit, OnDestroy {
     }
   }
 
+  discardCurrentCapture() {
+    this.mode$.next('capture');
+    this.startPreviewCamera();
+    this.removeCurrentCapture();
+  }
+
+  async confirmCurrentCapture() {
+    if (this.curCaptureFilePath && this.curCaptureType) {
+      await this.customCameraService.uploadToCapture(
+        this.curCaptureFilePath,
+        this.curCaptureType
+      );
+      this.removeCurrentCapture();
+    }
+    this.leaveCustomCamera();
+  }
+
   async leaveCustomCamera() {
     return this.location.back();
   }
@@ -140,6 +176,13 @@ export class CustomCameraPage implements OnInit, OnDestroy {
     this.router.navigate(['/settings', 'go-pro', 'media-list-on-camera'], {
       state: { shouldStartPreviewCameraOnLeave: true },
     });
+  }
+
+  private removeCurrentCapture() {
+    this.customCameraService.removeFile(this.curCaptureFilePath);
+    this.curCaptureFilePath = undefined;
+    this.curCaptureMimeType = undefined;
+    this.curCaptureSrc = undefined;
   }
 
   // eslint-disable-next-line class-methods-use-this