+++ src/app/features/home/custom-camera/pre-publish-mode/pre-publish-mode.component.spec.ts
@@ -1,0 +1,26 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { SharedTestingModule } from '../../../../shared/shared-testing.module';
+
+import { PrePublishModeComponent } from './pre-publish-mode.component';
+
+describe('PrePublishModeComponent', () => {
+  let component: PrePublishModeComponent;
+  let fixture: ComponentFixture<PrePublishModeComponent>;
+
+  beforeEach(
+    waitForAsync(() => {
+      TestBed.configureTestingModule({
+        declarations: [PrePublishModeComponent],
+        imports: [SharedTestingModule],
+      }).compileComponents();
+
+      fixture = TestBed.createComponent(PrePublishModeComponent);
+      component = fixture.componentInstance;
+      fixture.detectChanges();
+    })
+  );
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});